Revert "Fix(Ergo Chain): Run black formatter"
This reverts commit c1cce789ef1a42f2e8fc1d308db13a65b6593e44.

+++ scripts/utility/microprice_calculator.py
@@ -22,7 +22,7 @@ class MicropricePMM(ScriptStrategyBase):
 
     # ! Advanced configuration variables
     show_data = False  # ? Controls whether current df is shown in status
-    path_to_data = "./data"  # ? Default file format './data/microprice_{trading_pair}_{exchange}_{date}.csv'
+    path_to_data = './data'  # ? Default file format './data/microprice_{trading_pair}_{exchange}_{date}.csv'
     interval_to_write = 60
     price_line_width = 60
     precision = 4  # ? should be the length of the ticksize
@@ -30,7 +30,7 @@ class MicropricePMM(ScriptStrategyBase):
     day_offset = 1  # ? How many days back to start looking for csv files to load data from
 
     # ! Script variabes
-    columns = ["date", "time", "bid", "bs", "ask", "as"]
+    columns = ['date', 'time', 'bid', 'bs', 'ask', 'as']
     current_dataframe = pd.DataFrame(columns=columns)
     time_to_write = 0
     markets = {exchange: {trading_pair}}
@@ -48,13 +48,13 @@ class MicropricePMM(ScriptStrategyBase):
             self.dump_data()
 
     def format_status(self) -> str:
-        bid, ask = itemgetter("bid", "ask")(self.get_bid_ask())
-        bar = "=" * self.price_line_width + "\n"
-        header = f"Trading pair: {self.trading_pair}\nExchange: {self.exchange}\n"
-        price_line = f"Adjusted Midprice: {self.compute_adjusted_midprice()}\n         Midprice: {round((bid + ask) / 2, 8)}\n                 = {round(self.compute_adjusted_midprice() - ((bid + ask) / 2), 20)}\n\n{self.get_price_line()}\n"
-        imbalance_line = f"Imbalance: {self.compute_imbalance()}\n{self.get_imbalance_line()}\n"
-        data = f"Data path: {self.get_csv_path()}\n"
-        g_star = f"g_star:\n{self.g_star}" if self.g_star is not None else ""
+        bid, ask = itemgetter('bid', 'ask')(self.get_bid_ask())
+        bar = '=' * self.price_line_width + '\n'
+        header = f'Trading pair: {self.trading_pair}\nExchange: {self.exchange}\n'
+        price_line = f'Adjusted Midprice: {self.compute_adjusted_midprice()}\n         Midprice: {round((bid + ask) / 2, 8)}\n                 = {round(self.compute_adjusted_midprice() - ((bid + ask) / 2), 20)}\n\n{self.get_price_line()}\n'
+        imbalance_line = f'Imbalance: {self.compute_imbalance()}\n{self.get_imbalance_line()}\n'
+        data = f'Data path: {self.get_csv_path()}\n'
+        g_star = f'g_star:\n{self.g_star}' if self.g_star is not None else ''
 
         return f"\n\n\n{bar}\n\n{header}\n{price_line}\n\n{imbalance_line}\nn_spread: {self.n_spread} {'tick' if self.n_spread == 1 else 'ticks'}\n\n\n{g_star}\n\n{data}\n\n{bar}\n\n\n"
 
@@ -63,15 +63,13 @@ class MicropricePMM(ScriptStrategyBase):
     # Every 'time_to_write' ticks, writes the dataframe to a csv file
     def record_data(self):
         # Fetch bid and ask data
-        bid, ask, bid_volume, ask_volume = itemgetter("bid", "ask", "bs", "as")(self.get_bid_ask())
+        bid, ask, bid_volume, ask_volume = itemgetter('bid', 'ask', 'bs', 'as')(self.get_bid_ask())
         # Fetch date and time in seconds
         date = datetime.datetime.now().strftime("%Y-%m-%d")
         time = self.current_timestamp
 
         data = [[date, time, bid, bid_volume, ask, ask_volume]]
-        self.current_dataframe = self.current_dataframe.append(
-            pd.DataFrame(data, columns=self.columns), ignore_index=True
-        )
+        self.current_dataframe = self.current_dataframe.append(pd.DataFrame(data, columns=self.columns), ignore_index=True)
         return
 
     def dump_data(self):
@@ -83,23 +81,23 @@ class MicropricePMM(ScriptStrategyBase):
             data = pd.read_csv(csv_path, index_col=[0])
         except Exception as e:
             self.logger().info(e)
-            self.logger().info(f"Creating new csv file at {csv_path}")
+            self.logger().info(f'Creating new csv file at {csv_path}')
             data = pd.DataFrame(columns=self.columns)
 
-        data = data.append(self.current_dataframe.iloc[: -self.range_of_imbalance], ignore_index=True)
+        data = data.append(self.current_dataframe.iloc[:-self.range_of_imbalance], ignore_index=True)
         data.to_csv(csv_path)
-        self.current_dataframe = self.current_dataframe.iloc[-self.range_of_imbalance :]
+        self.current_dataframe = self.current_dataframe.iloc[-self.range_of_imbalance:]
         return
 
-    # ! Data methods
+# ! Data methods
     def get_csv_path(self):
         # Get all files in self.path_to_data directory
         files = os.listdir(self.path_to_data)
         for i in files:
-            if i.startswith(f"microprice_{self.trading_pair}_{self.exchange}"):
-                len_data = len(pd.read_csv(f"{self.path_to_data}/{i}", index_col=[0]))
+            if i.startswith(f'microprice_{self.trading_pair}_{self.exchange}'):
+                len_data = len(pd.read_csv(f'{self.path_to_data}/{i}', index_col=[0]))
                 if len_data > self.data_size_min:
-                    return f"{self.path_to_data}/{i}"
+                    return f'{self.path_to_data}/{i}'
 
         # Otherwise just return today's file
         return f'{self.path_to_data}/microprice_{self.trading_pair}_{self.exchange}_{datetime.datetime.now().strftime("%Y-%m-%d")}.csv'
@@ -111,7 +109,7 @@ class MicropricePMM(ScriptStrategyBase):
         ask_volume = asks.iloc[0].amount
         best_bid = bids.iloc[0].price
         bid_volume = bids.iloc[0].amount
-        return {"bid": best_bid, "ask": best_ask, "bs": bid_volume, "as": ask_volume}
+        return {'bid': best_bid, 'ask': best_ask, 'bs': bid_volume, 'as': ask_volume}
 
     # ! Microprice methods
     def compute_adjusted_midprice(self):
@@ -126,7 +124,7 @@ class MicropricePMM(ScriptStrategyBase):
             self.g_star = g_star
             self.ticksize = ticksize
         # Compute adjusted midprice from G_star and mid
-        bid, ask = itemgetter("bid", "ask")(self.get_bid_ask())
+        bid, ask = itemgetter('bid', 'ask')(self.get_bid_ask())
         mid = (bid + ask) / 2
         G_star = self.g_star
         ticksize = self.ticksize
@@ -134,12 +132,12 @@ class MicropricePMM(ScriptStrategyBase):
 
         # ? Compute adjusted midprice
         last_row = self.current_dataframe.iloc[-1]
-        imb = last_row["bs"].astype(float) / (last_row["bs"].astype(float) + last_row["as"].astype(float))
+        imb = last_row['bs'].astype(float) / (last_row['bs'].astype(float) + last_row['as'].astype(float))
         # Compute bucket of imbalance
         imb_bucket = [abs(x - imb) for x in G_star.columns].index(min([abs(x - imb) for x in G_star.columns]))
         # Compute and round spread index to nearest ticksize
         spreads = G_star[G_star.columns[imb_bucket]].values
-        spread = last_row["ask"].astype(float) - last_row["bid"].astype(float)
+        spread = last_row['ask'].astype(float) - last_row['bid'].astype(float)
         # ? Generally we expect this value to be < self._n_spread so we log when it's > self._n_spread
         spread_bucket = round(spread / ticksize) * ticksize // ticksize - 1
         if spread_bucket >= n_spread:
@@ -178,43 +176,43 @@ class MicropricePMM(ScriptStrategyBase):
         return False
 
     def estimate(self, T, n_spread, n_imb):
-        no_move = T[T["dM"] == 0]
-        no_move_counts = no_move.pivot_table(
-            index=["next_imb_bucket"], columns=["spread", "imb_bucket"], values="time", fill_value=0, aggfunc="count"
-        ).unstack()
-        Q_counts = np.resize(np.array(no_move_counts[0 : (n_imb * n_imb)]), (n_imb, n_imb))
+        no_move = T[T['dM'] == 0]
+        no_move_counts = no_move.pivot_table(index=['next_imb_bucket'],
+                                             columns=['spread', 'imb_bucket'],
+                                             values='time',
+                                             fill_value=0,
+                                             aggfunc='count').unstack()
+        Q_counts = np.resize(np.array(no_move_counts[0:(n_imb * n_imb)]), (n_imb, n_imb))
         # loop over all spreads and add block matrices
         for i in range(1, n_spread):
-            Qi = np.resize(np.array(no_move_counts[(i * n_imb * n_imb) : (i + 1) * (n_imb * n_imb)]), (n_imb, n_imb))
+            Qi = np.resize(np.array(no_move_counts[(i * n_imb * n_imb):(i + 1) * (n_imb * n_imb)]), (n_imb, n_imb))
             Q_counts = block_diag(Q_counts, Qi)
-        move_counts = (
-            T[(T["dM"] != 0)]
-            .pivot_table(index=["dM"], columns=["spread", "imb_bucket"], values="time", fill_value=0, aggfunc="count")
-            .unstack()
-        )
+        move_counts = T[(T['dM'] != 0)].pivot_table(index=['dM'],
+                                                    columns=['spread', 'imb_bucket'],
+                                                    values='time',
+                                                    fill_value=0,
+                                                    aggfunc='count').unstack()
 
         R_counts = np.resize(np.array(move_counts), (n_imb * n_spread, 4))
         T1 = np.concatenate((Q_counts, R_counts), axis=1).astype(float)
         for i in range(0, n_imb * n_spread):
             T1[i] = T1[i] / T1[i].sum()
-        Q = T1[:, 0 : (n_imb * n_spread)]
-        R1 = T1[:, (n_imb * n_spread) :]
+        Q = T1[:, 0:(n_imb * n_spread)]
+        R1 = T1[:, (n_imb * n_spread):]
 
         K = np.array([-0.01, -0.005, 0.005, 0.01])
-        move_counts = T[(T["dM"] != 0)].pivot_table(
-            index=["spread", "imb_bucket"],
-            columns=["next_spread", "next_imb_bucket"],
-            values="time",
-            fill_value=0,
-            aggfunc="count",
-        )
+        move_counts = T[(T['dM'] != 0)].pivot_table(index=['spread', 'imb_bucket'],
+                                                    columns=['next_spread', 'next_imb_bucket'],
+                                                    values='time',
+                                                    fill_value=0,
+                                                    aggfunc='count')
 
         R2_counts = np.resize(np.array(move_counts), (n_imb * n_spread, n_imb * n_spread))
         T2 = np.concatenate((Q_counts, R2_counts), axis=1).astype(float)
 
         for i in range(0, n_imb * n_spread):
             T2[i] = T2[i] / T2[i].sum()
-        R2 = T2[:, (n_imb * n_spread) :]
+        R2 = T2[:, (n_imb * n_spread):]
         G1 = np.dot(np.dot(np.linalg.inv(np.eye(n_imb * n_spread) - Q), R1), K)
         B = np.dot(np.linalg.inv(np.eye(n_imb * n_spread) - Q), R2)
         return G1, B
@@ -231,27 +229,27 @@ class MicropricePMM(ScriptStrategyBase):
         ticksize = np.round(min(spread.loc[spread > 0]) * 100) / 100
         # T.spread=T.ask-T.bid
         # adds the spread and mid prices
-        T["spread"] = np.round((T["ask"] - T["bid"]) / ticksize) * ticksize
-        T["mid"] = (T["bid"] + T["ask"]) / 2
+        T['spread'] = np.round((T['ask'] - T['bid']) / ticksize) * ticksize
+        T['mid'] = (T['bid'] + T['ask']) / 2
         # filter out spreads >= n_spread
         T = T.loc[(T.spread <= n_spread * ticksize) & (T.spread > 0)]
-        T["imb"] = T["bs"] / (T["bs"] + T["as"])
+        T['imb'] = T['bs'] / (T['bs'] + T['as'])
         # discretize imbalance into percentiles
-        T["imb_bucket"] = pd.qcut(T["imb"], n_imb, labels=False, duplicates="drop")
-        T["next_mid"] = T["mid"].shift(-dt)
+        T['imb_bucket'] = pd.qcut(T['imb'], n_imb, labels=False, duplicates='drop')
+        T['next_mid'] = T['mid'].shift(-dt)
         # step ahead state variables
-        T["next_spread"] = T["spread"].shift(-dt)
-        T["next_time"] = T["time"].shift(-dt)
-        T["next_imb_bucket"] = T["imb_bucket"].shift(-dt)
+        T['next_spread'] = T['spread'].shift(-dt)
+        T['next_time'] = T['time'].shift(-dt)
+        T['next_imb_bucket'] = T['imb_bucket'].shift(-dt)
         # step ahead change in price
-        T["dM"] = np.round((T["next_mid"] - T["mid"]) / ticksize * 2) * ticksize / 2
+        T['dM'] = np.round((T['next_mid'] - T['mid']) / ticksize * 2) * ticksize / 2
         T = T.loc[(T.dM <= ticksize * 1.1) & (T.dM >= -ticksize * 1.1)]
         # symetrize data
         T2 = T.copy(deep=True)
-        T2["imb_bucket"] = n_imb - 1 - T2["imb_bucket"]
-        T2["next_imb_bucket"] = n_imb - 1 - T2["next_imb_bucket"]
-        T2["dM"] = -T2["dM"]
-        T2["mid"] = -T2["mid"]
+        T2['imb_bucket'] = n_imb - 1 - T2['imb_bucket']
+        T2['next_imb_bucket'] = n_imb - 1 - T2['next_imb_bucket']
+        T2['dM'] = -T2['dM']
+        T2['mid'] = -T2['mid']
         T3 = pd.concat([T, T2])
         T3.index = pd.RangeIndex(len(T3.index))
         return T3, ticksize
@@ -265,49 +263,43 @@ class MicropricePMM(ScriptStrategyBase):
             self.logger().info(e)
             df = self.current_dataframe
 
-        df["time"] = df["time"].astype(float)
-        df["bid"] = df["bid"].astype(float)
-        df["ask"] = df["ask"].astype(float)
-        df["bs"] = df["bs"].astype(float)
-        df["as"] = df["as"].astype(float)
-        df["mid"] = (df["bid"] + df["ask"]) / float(2)
-        df["imb"] = df["bs"] / (df["bs"] + df["as"])
+        df['time'] = df['time'].astype(float)
+        df['bid'] = df['bid'].astype(float)
+        df['ask'] = df['ask'].astype(float)
+        df['bs'] = df['bs'].astype(float)
+        df['as'] = df['as'].astype(float)
+        df['mid'] = (df['bid'] + df['ask']) / float(2)
+        df['imb'] = df['bs'] / (df['bs'] + df['as'])
         return df
 
     def compute_imbalance(self) -> Decimal:
         if self.get_df().empty or self.current_dataframe.empty:
-            self.logger().info("No data to compute imbalance, recording data")
+            self.logger().info('No data to compute imbalance, recording data')
             self.recording_data = True
             return Decimal(-1)
-        bid_size = self.current_dataframe["bs"].sum()
-        ask_size = self.current_dataframe["as"].sum()
+        bid_size = self.current_dataframe['bs'].sum()
+        ask_size = self.current_dataframe['as'].sum()
         return round(Decimal(bid_size) / Decimal(bid_size + ask_size), self.precision * 2)
 
     # ! Format status methods
     def get_price_line(self) -> str:
         # Get best bid and ask
-        bid, ask = itemgetter("bid", "ask")(self.get_bid_ask())
+        bid, ask = itemgetter('bid', 'ask')(self.get_bid_ask())
         # Mid price is center of line
-        price_line = int(self.price_line_width / 2) * "-" + "|" + int(self.price_line_width / 2) * "-"
+        price_line = int(self.price_line_width / 2) * '-' + '|' + int(self.price_line_width / 2) * '-'
         # Add bid, adjusted midprice,
         bid_offset = int(self.price_line_width / 2 - len(str(bid)) - (len(str(self.compute_adjusted_midprice())) / 2))
         ask_offset = int(self.price_line_width / 2 - len(str(ask)) - (len(str(self.compute_adjusted_midprice())) / 2))
-        labels = (
-            str(bid) + bid_offset * " " + str(self.compute_adjusted_midprice()) + ask_offset * " " + str(ask) + "\n"
-        )
+        labels = str(bid) + bid_offset * ' ' + str(self.compute_adjusted_midprice()) + ask_offset * ' ' + str(ask) + '\n'
         # Create microprice of size 'price_line_width' with ends best bid and ask
         mid = (bid + ask) / 2
         spread = ask - bid
         microprice_adjustment = self.compute_adjusted_midprice() - mid + (spread / 2)
         adjusted_midprice_i = int(microprice_adjustment / spread * self.price_line_width) + 1
-        price_line = price_line[:adjusted_midprice_i] + "m" + price_line[adjusted_midprice_i:]
+        price_line = price_line[:adjusted_midprice_i] + 'm' + price_line[adjusted_midprice_i:]
         return labels + price_line
 
     def get_imbalance_line(self) -> str:
-        imb_line = int(self.price_line_width / 2) * "-" + "|" + int(self.price_line_width / 2) * "-"
-        imb_line = (
-            imb_line[: int(self.compute_imbalance() * self.price_line_width)]
-            + "i"
-            + imb_line[int(self.compute_imbalance() * self.price_line_width) :]
-        )
+        imb_line = int(self.price_line_width / 2) * '-' + '|' + int(self.price_line_width / 2) * '-'
+        imb_line = imb_line[:int(self.compute_imbalance() * self.price_line_width)] + 'i' + imb_line[int(self.compute_imbalance() * self.price_line_width):]
         return imb_line
